+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.0.0-beta.69
+
+### instant-apps-control-panel
+
+- Component to group components together.
+
+## v1.0.0-beta.68
+
 ## v1.0.0-beta.67
 
 ### instant-apps-filter-list